chatbox-course/ThinkRook-Chatbot/requirements.txt
streamlit==1.50.0
google-genai==1.39.1
google-generativeai==0.8.5
pandas==2.3.2
pillow==11.3.0
ollama==0.6.0
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5